Fix more regressions
